installed npm package to help with Gmaps
