+++ package.json
@@ -23,6 +23,7 @@
     "classnames": "^2.2.6",
     "dotenv": "^6.1.0",
     "express": "^4.16.4",
+    "google-map-react": "^1.0.9",
     "hbs": "^4.0.1",
     "node": "^10.12.0",
     "pg-promise": "^8.5.1",
